m: don't use autoreconf, patch Makefile.in files
Currently Makefile.am files in examples/ are
edited with sed (REINPLACE_CMD) in the post-patch
target. This requires USES+=autoreconf
(to run ./autogen.sh).

Make the same changes to Makefile.in files
instead. We can do this now that we download
the correct tarball, lepton-eda-1.9.13.tar.gz,
and not _GH0 stuff, which is a file listed as
"Source code (tar.gz)" on the lepton-eda 1.9.13
release page on github (url:
https://github.com/lepton-eda/lepton-eda/releases/tag/1.9.13-20201211).